normi final
